% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunction.R
\name{conjunction}
\alias{conjunction}
\title{Conjunction}
\usage{
conjunction(name, date, social)
}
\arguments{
\item{name}{A string containing the name of a person}

\item{date}{A string containing the date of birth of in the format of Day/Month/Year}

\item{social}{A numeric variable containing the social security number of user}
}
\description{
Conjunction
}
\examples{
conjunction("Melanie", "??/??/????", 123456789)
}
